+++ CHANGELOG.md
@@ -18,6 +18,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Added quantize method to bounding boxes which pushes the left-top coordinates
   to their floor and the right-bottom coordinates to their ceiling.
 
+* `Detections.draw_on` can now accept color='classes'
+
 ### Fixed
 * `kwimage.Detections` now correctly handles `None` data values. Previously
   `None` was converted to an `array(None, dtype=object)`.